Fix issue with guild detection
Discord apparently doesn't send all the info in the fetch message endpoint as given on a message event, resulting in the need to fetch the message's channel and use that to determine if the message is in a guild or not. This is because the channel fetch endpoint actually returns the information necessary to tell if it is in a guild or not.